ewis/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
google-generativeai==0.5.4
httpx==0.27.0
sse-starlette==2.1.0
alembic==1.13.1
asyncpg==0.29.0
markdown==3.6
pdfminer.six==20231228
